splitted recent documents & tag cloud into separated components
